checker(#8722): a wide literal in an unannotated tuple / array init settles at i64
An unannotated binding settles a still-polymorphic integer init at i64 when a
literal in it has no i32 reading (#3676, #8668), but that walk only reached a
SCALAR init. A tuple or array init is not itself a numeric expression, so
`var t = (1, 4611686018427387904)` and `var xs = [4611686018427387904]` left
their elements polymorphic and every native backend lowered them at the i32
default — 2^62 reads back as 0 — while `-interp` computed the full value, with
no diagnostic either way. The self-host types such an element by magnitude and
gets it right, so this was a native-vs-self-host value divergence as well as a
silent wrong answer.

widenCompositeByLiterals rewrites the polymorphic integer slots of a tuple /
array / slice init to i64 where the matching element tree holds such a
literal, and settleNumeric then settles the init at the rewritten type. It
descends nested composites and the arms of an if / match, and leaves a slot
whose elements all read at i32 polymorphic, so `var t = (1, 2)` is unchanged.
An element past u64 is now refused against the i64 it settles at rather than
the i32 it never reached, which is what the two moved rows in
int_literal_range_test record.

Blast radius, measured against a compiler built from the merge-base: the 218
Fern sources in the tree (examples, coreutils, the self-host tree, std + core)
check to the same verdict; the 42 of them that are whole programs compile to
byte-identical x86-64 binaries; and all 570 conformance cases produce
identical output and exit codes under -interp. Nothing in the tree can move:
across those 788 files there is no unannotated tuple / array init holding a
decimal literal past i32 at all.

+++ internal/checker/int_literal_range_test.go
@@ -211,7 +211,8 @@ func checkErrors(t *testing.T, src string) []*Error {
 // own column — the report every other over-range literal gets (#8563). With no
 // context to name, the refusal names the type the literal would otherwise
 // default to: i64 for an unannotated binding (the #3676 widening, which reads
-// through the binding's arithmetic too), i32 anywhere else.
+// through the binding's arithmetic and its tuple / array elements too), i32
+// anywhere else.
 func TestIntLiteralPastU64IsE047(t *testing.T) {
 	const lit = "18446744073709551616"
 	cases := []struct {
@@ -233,9 +234,10 @@ func TestIntLiteralPastU64IsE047(t *testing.T) {
 		// A comparison of two literals settles them itself, at the i64 the
 		// wide one selects (#8668).
 		{`var b: boolean = ` + lit + ` > 1;`, lit, "i64", 41},
-		// Nothing settles these; they would lower as the i32 default.
-		{`var t = (` + lit + `, 1);`, lit, "i32", 33},
-		{`var xs = [` + lit + `];`, lit, "i32", 34},
+		// A composite init settles its elements at the same i64 the wide
+		// literal selects for a scalar one (#8722).
+		{`var t = (` + lit + `, 1);`, lit, "i64", 33},
+		{`var xs = [` + lit + `];`, lit, "i64", 34},
 		// Hex is quoted as written.
 		{`var a: u64 = 0xFFFFFFFFFFFFFFFFFFFF;`, "0xFFFFFFFFFFFFFFFFFFFF", "u64", 37},
 		// Float context has no integer to promote.
@@ -267,10 +269,11 @@ func TestIntLiteralPastU64IsE047(t *testing.T) {
 // A literal past i64 max fits only u64 (or i64 as its minimum), so it is valid
 // only where a context settles it. One left unsettled — a tuple or array
 // element — lowered as the i32 default and wrapped silently (the #8449
-// family). It is refused against that default now; a settled one is untouched.
-// An unannotated binding's arithmetic and a comparison of two literals settle
-// themselves at the i64 default a wide literal selects (#8668), so the literal
-// is judged there.
+// family). An unannotated binding's arithmetic, a comparison of two literals
+// (#8668) and a tuple / array element (#8722) now settle themselves at the i64
+// default a wide literal selects, so the literal is judged there; one nothing
+// settles at all is refused against the i32 default, and a settled one is
+// untouched.
 func TestWideIntLiteralLeftUnsettledIsE047(t *testing.T) {
 	rejected := []struct {
 		src     string
@@ -278,8 +281,8 @@ func TestWideIntLiteralLeftUnsettledIsE047(t *testing.T) {
 		wantIn  string
 	}{
 		{`var b = 9223372036854775808 + 1;`, 32, "i64"},
-		{`var t = (9223372036854775808, 1);`, 33, "i32"},
-		{`var xs = [18446744073709551615];`, 34, "i32"},
+		{`var t = (9223372036854775808, 1);`, 33, "i64"},
+		{`var xs = [18446744073709551615];`, 34, "i64"},
 		{`var b: boolean = 9223372036854775808 > 1;`, 41, "i64"},
 	}
 	for _, c := range rejected {
